chore(matrix): remove obsolete intermediate MatrixNode class
